// src/models/view_models/event_detail.ts
export interface EventDetail {
    event_id: string;
    title: string;
    image_url: string;
    line1: string;
    line2: string;
    city: string;
    state: string;
    zip: string;
    min_age: number;
    max_guests: number;
    scene: string;
    description: string;
    artists: string;
    promoter: string;
}

// - event_id
// - title
// - image_url
// - line1
// - line2
// - city
// - state
// - zip
// - min_age
// - max_guests
// - scene
// - description
// - artists
// - promoter